** Moved ui specific test methods into new UITest class from IntegrationTest class
